+++ packages/vm/src/lib.rs
@@ -30,6 +30,5 @@ pub use crate::errors::{
 pub use crate::features::features_from_csv;
 pub use crate::ffi::{FfiError, FfiResult, GasInfo};
 pub use crate::instance::{GasReport, Instance, InstanceOptions};
-pub use crate::modules::FileSystemCache;
 pub use crate::serde::{from_slice, to_vec};
 pub use crate::traits::{Api, Extern, Querier, Storage};