+++ packages/vm/src/modules/mod.rs
@@ -1,0 +1,3 @@
+mod file_system_cache;
+
+pub use file_system_cache::FileSystemCache;